+++ pkg/analyzer_plugin/lib/utilities/range_factory.dart
@@ -63,11 +63,15 @@ class RangeFactory {
 
   /// Return the deletion range of the [node], considering the spaces and
   /// comments before and after it.
-  SourceRange deletionRange(AstNode node) {
+  ///
+  /// If a non-`null` [overrideEnd] is supplied, it will be used in place of
+  /// [AstNode.endToken] to determine the range of tokens to delete.
+  SourceRange deletionRange(AstNode node, {Token? overrideEnd}) {
     var begin = node.beginToken;
     begin = begin.precedingComments ?? begin;
 
-    var end = node.endToken.next!;
+    var initialEndToken = overrideEnd ?? node.endToken;
+    var end = initialEndToken.next!;
     end = end.precedingComments ?? end;
 
     int startOffset;
@@ -84,7 +88,7 @@ class RangeFactory {
       } else {
         startOffset = begin.previous!.end;
       }
-      endOffset = node.endToken.end;
+      endOffset = initialEndToken.end;
     } else {
       startOffset = begin.offset;
       endOffset = end.offset;