+++ pkg/vm_service/lib/src/vm_service.dart
@@ -1737,7 +1737,7 @@ class VmService {
     _disposed = true;
     await _streamSub.cancel();
     _outstandingRequests.forEach((id, request) {
-      request._completer.completeError(RPCError(
+      request.completeError(RPCError(
         request.method,
         RPCErrorKind.kServerError.code,
         'Service connection disposed',