+++ runtime/vm/service.cc
@@ -52,7 +52,6 @@
 #include "vm/symbols.h"
 #include "vm/timeline.h"
 #include "vm/version.h"
-#include "vm/visitor.h"
 
 #if defined(SUPPORT_PERFETTO)
 #include "vm/perfetto_utils.h"
@@ -364,34 +363,28 @@ static const char* GetVMName() {
   return vm_name;
 }
 
-ServiceIdZone::ServiceIdZone(ObjectIdRing::IdPolicy policy) : policy_(policy) {}
+ServiceIdZone::ServiceIdZone() {}
 
 ServiceIdZone::~ServiceIdZone() {}
 
-RingServiceIdZone::RingServiceIdZone(ObjectIdRing::IdPolicy policy)
-    : ServiceIdZone(policy), ring_() {}
+RingServiceIdZone::RingServiceIdZone()
+    : ring_(nullptr), policy_(ObjectIdRing::kAllocateId) {}
 
 RingServiceIdZone::~RingServiceIdZone() {}
 
-int32_t RingServiceIdZone::GetIdForObject(const ObjectPtr obj) {
-  return ring_.GetIdForObject(obj, policy());
-}
-
-ObjectPtr RingServiceIdZone::GetObjectForId(int32_t id,
-                                            ObjectIdRing::LookupResult* kind) {
-  return ring_.GetObjectForId(id, kind);
+void RingServiceIdZone::Init(ObjectIdRing* ring,
+                             ObjectIdRing::IdPolicy policy) {
+  ring_ = ring;
+  policy_ = policy;
 }
 
 char* RingServiceIdZone::GetServiceId(const Object& obj) {
+  ASSERT(ring_ != nullptr);
   Thread* thread = Thread::Current();
   Zone* zone = thread->zone();
   ASSERT(zone != nullptr);
-  const intptr_t id = GetIdForObject(obj.ptr());
+  const intptr_t id = ring_->GetIdForObject(obj.ptr(), policy_);
   return zone->PrintToString("objects/%" Pd "", id);
-}
-
-void RingServiceIdZone::VisitPointers(ObjectPointerVisitor& visitor) const {
-  ring_.VisitPointers(&visitor);
 }
 
 // TODO(johnmccutchan): Unify embedder service handler lists and their APIs.
@@ -1005,11 +998,24 @@ ErrorPtr Service::InvokeMethod(Isolate* I,
     const char* id_zone_param = js.LookupParam("_idZone");
 
     if (id_zone_param != nullptr) {
-      // TODO(derekxu16): Support creating, deleting, and selecting custom
-      // Service ID zones. For now, always return an error.
-      PrintInvalidParamError(&js, "_idZone");
-      js.PostReply();
-      return T->StealStickyError();
+      // Override id zone.
+      if (strcmp("default", id_zone_param) == 0) {
+        // Ring with eager id allocation. This is the default ring and default
+        // policy.
+        // Nothing to do.
+      } else if (strcmp("default.reuse", id_zone_param) == 0) {
+        // Change the default ring's policy.
+        RingServiceIdZone* zone =
+            reinterpret_cast<RingServiceIdZone*>(js.id_zone());
+        zone->set_policy(ObjectIdRing::kReuseId);
+      } else {
+        // TODO(johnmccutchan): Support creating, deleting, and selecting
+        // custom service id zones.
+        // For now, always return an error.
+        PrintInvalidParamError(&js, "_idZone");
+        js.PostReply();
+        return T->StealStickyError();
+      }
     }
     const char* c_method_name = method_name.ToCString();
 
@@ -1830,12 +1836,13 @@ static ObjectPtr LookupObjectId(Thread* thread,
     return Object::null();
   }
 
+  ObjectIdRing* ring = thread->isolate()->EnsureObjectIdRing();
   intptr_t id = -1;
   if (!GetIntegerId(arg, &id)) {
     *kind = ObjectIdRing::kInvalid;
     return Object::null();
   }
-  return thread->isolate()->GetDefaultServiceIdZone().GetObjectForId(id, kind);
+  return ring->GetObjectForId(id, kind);
 }
 
 static ObjectPtr LookupClassMembers(Thread* thread,