+++ runtime/vm/service.h
@@ -6,7 +6,6 @@
 #define RUNTIME_VM_SERVICE_H_
 
 #include <atomic>
-#include <memory>
 
 #include "include/dart_tools_api.h"
 
@@ -37,22 +36,13 @@ class String;
 
 class ServiceIdZone {
  public:
-  explicit ServiceIdZone(ObjectIdRing::IdPolicy policy);
+  ServiceIdZone();
   virtual ~ServiceIdZone();
 
-  ObjectIdRing::IdPolicy policy() const { return policy_; }
-
-  virtual int32_t GetIdForObject(const ObjectPtr obj) = 0;
-  virtual ObjectPtr GetObjectForId(int32_t id,
-                                   ObjectIdRing::LookupResult* kind) = 0;
   // Returned string will be zone allocated.
   virtual char* GetServiceId(const Object& obj) = 0;
-  virtual void VisitPointers(ObjectPointerVisitor& visitor) const = 0;
 
  private:
-  ObjectIdRing::IdPolicy policy_;
-
-  friend class ServiceIdZonePolicyOverrideScope;
 };
 
 #define ISOLATE_SERVICE_ID_FORMAT_STRING "isolates/%" Pd64 ""
@@ -60,19 +50,23 @@ class ServiceIdZone {
 #define ISOLATE_GROUP_SERVICE_ID_FORMAT_STRING                                 \
   ISOLATE_GROUP_SERVICE_ID_PREFIX "%" Pu64 ""
 
-class RingServiceIdZone final : public ServiceIdZone {
+class RingServiceIdZone : public ServiceIdZone {
  public:
-  explicit RingServiceIdZone(ObjectIdRing::IdPolicy policy);
-  ~RingServiceIdZone() final;
+  RingServiceIdZone();
+  virtual ~RingServiceIdZone();
 
-  int32_t GetIdForObject(const ObjectPtr obj) final;
-  ObjectPtr GetObjectForId(int32_t id, ObjectIdRing::LookupResult* kind) final;
+  void Init(ObjectIdRing* ring, ObjectIdRing::IdPolicy policy);
+
   // Returned string will be zone allocated.
-  char* GetServiceId(const Object& obj) final;
-  void VisitPointers(ObjectPointerVisitor& visitor) const final;
+  virtual char* GetServiceId(const Object& obj);
+
+  void set_policy(ObjectIdRing::IdPolicy policy) { policy_ = policy; }
+
+  ObjectIdRing::IdPolicy policy() const { return policy_; }
 
  private:
-  ObjectIdRing ring_;
+  ObjectIdRing* ring_;
+  ObjectIdRing::IdPolicy policy_;
 };
 
 class StreamInfo {